services/payment-service/requirements.txt
Flask
stripe
psycopg2-binary
python-dotenv
